backend/requirements.txt
flask
flask-cors
requests
beautifulsoup4
selenium
webdriver-manager
fake-useragent
pymongo
python-dotenv
apscheduler
google-search-results
PyJWT
werkzeug
scikit-learn
numpy
pandas
gunicorn
Pillow